Moving Game to Pages instead of components
