+++ src/App.tsx
@@ -1,10 +1,10 @@
 import './App.css';
-import Home from './pages/Home';
+import Game from './pages/Game';
 
 function App() {
 	return (
 		<div className='App'>
-			<Home />
+			<Game />
 		</div>
 	);
 }
