+++ src/components/Board/index.tsx
@@ -6,11 +6,11 @@ import Ball from '../Ball';
 import Paddle from '../Paddle';
 import useKeyPress from '../../util/useKeyPress'; // Import the hook
 import AABBIntersect from '../../util/aabbIntersect';
-import getMachineSpeed from 'components/Game/getMachineSpeed';
+import getMachineSpeed from 'pages/Game/getMachineSpeed';
 import { PaddleType } from './types/PaddleType';
 import { BallPositionType } from './types/BallPositionType';
 import { BallVelocityType } from './types/BallVelocityType';
-import { OpponentMode } from 'components/Game/enums/OpponentMode';
+import { OpponentMode } from 'pages/Game/enums/OpponentMode';
 
 function Board(props: BoardProps) {
 	// Simple state variables
