+++ src/pages/Game/enums/Difficulty.ts
