+++ src/pages/Game/enums/GameMode.ts
