+++ src/pages/Game/getMachineSpeed.ts
