+++ src/pages/Game/types/GameDataType.ts
