+++ src/pages/Game/types/GameOptionsType.ts
