+++ src/pages/Game/types/GameSettingsType.ts
