+++ src/pages/Game/types/ScoreType.ts
