--- src/pages/Home/index.tsx
@@ -1,12 +1,0 @@
-import Game from '../../components/Game';
-import './styles.css';
-
-function Home() {
-	return (
-		<div className='home-container'>
-			<Game />
-		</div>
-	);
-}
-
-export default Home;
